requirements.txt
dnspython==2.3.0
pymongo==4.3.3
python-dotenv==1.0.0
